backend/requirements.txt
# requirements.txt - Python dependencies cho ML

# Data processing
pandas==2.1.1
numpy==1.26.2

# Machine Learning
scikit-learn==1.3.2
joblib==1.3.2

# API
flask==3.0.0
flask-cors==4.0.0

# Database
pymongo==4.6.0

# Environment
python-dotenv==1.0.0

# Optional: Performance
scipy==1.11.4
